refactor(sui-genesis-builder): split stardust::types module
