+++ crates/sui-genesis-builder/src/stardust/migration.rs
@@ -38,7 +38,7 @@ use sui_types::{
     MOVE_STDLIB_PACKAGE_ID, STARDUST_PACKAGE_ID, SUI_FRAMEWORK_PACKAGE_ID,
 };
 
-use super::types::OutputHeader;
+use super::types::snapshot::OutputHeader;
 use crate::process_package;
 
 /// The dependencies of the generated packages for native tokens.
