+++ crates/sui-genesis-builder/src/stardust/types/mod.rs
@@ -1,0 +1,1 @@
+pub mod snapshot;
